requirements.txt
python-telegram-bot==21.6
openai==1.40.0